Edited models and apis

+++ src/main/java/com/example/hbs/service/BookingService.java
@@ -3,6 +3,7 @@ package com.example.hbs.service;
 import com.example.hbs.model.Booking;
 import com.example.hbs.repository.BookingReporitory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -10,11 +11,17 @@ public class BookingService {
     @Autowired
     BookingReporitory bookingReporitory;
 
-    public void addBooking(Booking booking) {
+    public ResponseEntity<String> addBooking(Booking booking) {
         bookingReporitory.save(booking);
+        return ResponseEntity.ok("Booking Added");
     }
 
-    public void deleteBooking(Long id) {
-        bookingReporitory.deleteById(id);
+    public ResponseEntity<String> deleteBooking(Long id) {
+        if (bookingReporitory.findById(id).isPresent()) {
+            bookingReporitory.deleteById(id);
+            return ResponseEntity.ok("Booking Deleted");
+        } else {
+            return ResponseEntity.badRequest().body("No such Booking is present");
+        }
     }
 }
